server/requirements.txt
python-socketio
eventlet
pycryptodome